update intendation to two spaces
